+++ Diary_app/src/screens/DiaryEdit/FunctionComponents/FunctionComponents.tsx
@@ -1,18 +1,15 @@
 import { View, Text, Dimensions, TouchableOpacity, Animated } from 'react-native';
 import React, { useState, useEffect, useRef } from 'react';
-import { SafeAreaView } from 'react-native';
 import styled from 'styled-components';
 import { ScrollView, Image } from 'react-native';
-import { CameraOptions, launchCamera, launchImageLibrary } from 'react-native-image-picker';
-import { PermissionsAndroid } from 'react-native';
-import { TouchableHighlight } from 'react-native';
 import { TextInput } from 'react-native';
-import { useNavigation } from '@react-navigation/native';
-import { countEx, setCountEx, updateTagZoneEx } from '../DiaryEdit';
+import { countEx, setCountEx, enableDeleteEx, setEnableDeleteEx } from '../DiaryDetail';
+import { openCameraEx, openGalleryEx } from './PhotoZone';
+import { setWriteDiaryEx, writeDiaryEx } from './NoteZone';
 const icon_camera = require('../FunctionComponents/icons/camera.png');
 const icon_gallery = require('../FunctionComponents/icons/gallery.png');
 const icon_note = require('../FunctionComponents/icons/note.png');
-const icon_delete = require('../FunctionComponents/icons/delete.png');
+// const icon_delete = require('../FunctionComponents/icons/delete.png');
 const icon_edit = require('../FunctionComponents/icons/edit.png');
 const StyledHorizontallyAlignedItems = styled(View)`
 flex-direction: row;
@@ -20,7 +17,6 @@ flex-direction: row;
  align-items: center;
  flex:1;
  padding-horizontal: 20;
-
 `
 const StyledCircleButton = styled(TouchableOpacity)`
 border-width: 1;
@@ -38,16 +34,6 @@ text-align: center;
 color: white;
 `
 
-const StyledHorizontalScrollViewOverflow = styled(View)`
-overflow: hidden;
-border-radius:30;
-margin:5px;
-`
-const StyledScrollHorizontalScroll = styled(ScrollView)`
-background-color: #666666AA;
-border-radius:30;
-paddingHorizontal: 5px;
-`
 
 const StyledTextInput = styled(TextInput)`
 border-radius: 50;
@@ -56,149 +42,10 @@ border-width: 1;
 padding: 15px;
 color:white;
 `
-export let setCountInc:any;
-export let setCountDec:any; //setState
-export let editEnable:boolean;
-export let setEditEnable:any;
-
-function Photos(props: any & JSX.Element): JSX.Element {
-    return (
-        <StyledHorizontalScrollViewOverflow >
-            <StyledScrollHorizontalScroll horizontal={true} showsHorizontalScrollIndicator={false}>
-                {
-                    props.content
-                }
-            </StyledScrollHorizontalScroll>
-        </StyledHorizontalScrollViewOverflow>
-    )
-}
-function FunctionComponents(): [JSX.Element, JSX.Element, JSX.Element, string] {
-    let options: CameraOptions = {
-        saveToPhotos: true,
-        mediaType: 'photo'
-    }
-    const [photo, setPhoto] = useState<string[]>([])
-    const [photoContent, setPhotoContent] = useState<JSX.Element[]>()
-
-    const [enableDelete, setEnableDelete] = useState<boolean>(false)
-    // const [count, setCount] = useState<number>(0)
-    const count = countEx
-    const setCount = setCountEx
-    
-    useEffect(() => {
-        setEnableDelete(enableDelete)
-        setPhoto(photo)
-    })
-    const openCamera = async () => {
-        try {
-            const granted = await PermissionsAndroid.request(
-                PermissionsAndroid.PERMISSIONS.CAMERA,
-            );
-            if (granted === PermissionsAndroid.RESULTS.GRANTED) {
-                const result = await launchCamera(options)
-                if (result.assets !== undefined && result.assets[0].uri !== undefined) {
-                    photo.push(result.assets[0].uri)
-                    setPhoto(photo)
-                    setCount(count + 1)
-                    // increaseCount()
-                    setEnableDelete(false)
-                    updatePhotoContent()
-                }
-            }
-        } catch (error) {
-            console.log(error) // handle error
-        }
-        setVisible(false)
-        animationReset()
-
-    }
-    const openGallery = async () => {
-        const result = await launchImageLibrary(options)
-        if (result.assets !== undefined && result.assets[0].uri !== undefined) {
-            photo.push(result.assets[0].uri)
-            setPhoto(photo)
-            setCount(count + 1)
-            // increaseCount()
-            setEnableDelete(false)
-            updatePhotoContent()
-
-        }
-        setVisible(false)
-        animationReset()
-    }
-    const deletePhoto = (deletable: boolean, index: number) => {
-        if (deletable) {
-            // if (enableDelete)
-            delete photo[index]
-            setPhoto(photo)
-            setCount(count - 1)
-            // decreaseCount()
-            updatePhotoContent()
-
-            // setEnableDelete(false)
-        }
-        setVisible(false)
-        animationReset()
-    }
-    useEffect(() => {
-        if (enableDelete && count == 0) {
-            setEnableDelete(false)
-        }
-        updateTagZoneEx()
-        updatePhotoContent() //새 state 값과 함께 해당지역 새로 만들어주면됨~
-    }, [enableDelete, count])
-
-    //포토에 새로운값이 들어왔다면, 즉각 컨텐츠를 업데이트.
-    const updatePhotoContent = () => {
-        let contentHolder = (
-            photo.map((picture: string, index: number) => (
-                <View style={{ position: 'relative' }}>
-                    <TouchableHighlight onLongPress={() => setEnableDelete(true)} key={index}
-                        style={{
-                            margin: 5,
-
-                        }}>
-
-                        <Image source={{ uri: picture }} key={index} style={{ height: 150, width: 150, borderRadius: 50 }} />
-                    </TouchableHighlight >
-                    <View style={{ position: 'absolute', right: 0 }}>
-                    {
-                        enableDelete &&
-                        //must componentize this. (duplicate code in diaryedit)
-                        <TouchableHighlight
-                            onPress={() => {deletePhoto(enableDelete, index)}}
-                            style={{
-                                borderRadius: 50,
-                                borderColor: 'white',
-                                backgroundColor:'black',
-                                width:25,
-                                height:25,
-                            }}>
-                            <Text style={{
-                                color: 'white',
-                                textAlign:'center'
-                            }}>
-                                X
-                            </Text>
-                        </TouchableHighlight>
-                    }
-                    </View>
-                </View>
-            ))
-        )
-
-        setPhotoContent(contentHolder)
-    }
-
-    //remove pictures
-    const toggleDeletePhoto = () => {
-        setEnableDelete(!enableDelete)
-    }
-    const navigation = useNavigation();
-    const removeCurrentDiary = () => {
-        //remove algorithm is not implemented yet.
-        navigation.navigate('Diary')
-    }
+export let showPlusButtonEx:any;
+function FunctionComponents(): JSX.Element {
+    const enableDelete = enableDeleteEx
+    const setEnableDelete = setEnableDeleteEx
 
     const button_size = 25
     const start_x = 0
@@ -213,7 +60,8 @@ function FunctionComponents(): [JSX.Element, JSX.Element, JSX.Element, string] {
     const [animation4] = useState(new Animated.ValueXY({ x: start_x, y: start_y }));
     //edit
     const [animation5] = useState(new Animated.ValueXY({ x: start_x, y: start_y }));
-    
+
+    //the + button not visible by default
     const [visible, setVisible] = useState<boolean>(false)
     const moveComponent = () => {
         Animated.timing(animation1, {
@@ -242,57 +90,24 @@ function FunctionComponents(): [JSX.Element, JSX.Element, JSX.Element, string] {
             useNativeDriver: true, // Use native driver for better performance
         }).start();
     };
+    const setTransform = (x: Animated.Value, y: Animated.Value) => {
+        return{
+            transform: [
+                {
+                    translateX: x,
+                },
+                {
+                    translateY: y,
+                },
+            ]
+        }
+    }
 
-    const animatedStyle1 = {
-        transform: [
-            {
-                translateX: animation1.x,
-            },
-            {
-                translateY: animation1.y,
-            },
-        ],
-    };
-    const animatedStyle2 = {
-        transform: [
-            {
-                translateX: animation2.x,
-            },
-            {
-                translateY: animation2.y,
-            },
-        ],
-    };
-    const animatedStyle3 = {
-        transform: [
-            {
-                translateX: animation3.x,
-            },
-            {
-                translateY: animation3.y,
-            },
-        ],
-    };
-    const animatedStyle4 = {
-        transform: [
-            {
-                translateX: animation4.x,
-            },
-            {
-                translateY: animation4.y,
-            },
-        ],
-    };
-    const animatedStyle5 = {
-        transform: [
-            {
-                translateX: animation5.x,
-            },
-            {
-                translateY: animation5.y,
-            },
-        ],
-    };
+    const animatedStyle1 = setTransform(animation1.x, animation1.y)
+    const animatedStyle2 = setTransform(animation2.x, animation2.y)
+    const animatedStyle3 = setTransform(animation3.x, animation3.y)
+    const animatedStyle4 = setTransform(animation4.x, animation4.y)
+    const animatedStyle5 = setTransform(animation5.x, animation5.y)
     const animationReset = () => {
         animation1.setValue({ x: start_x, y: start_y });
         animation2.setValue({ x: start_x, y: start_y });
@@ -300,76 +115,25 @@ function FunctionComponents(): [JSX.Element, JSX.Element, JSX.Element, string] {
         animation4.setValue({ x: start_x, y: start_y });
         animation5.setValue({ x: start_x, y: start_y });
     }
-    const [writeDiary, setWriteDiary] = useState<boolean>(false)
-    const [text, setText] = useState<string>('')
-    const [returnText, setReturnText] = useState<string>('')
-
     
-    //export count variable to DiaryEditFile
-    
-    editEnable = enableDelete
-    setEditEnable = setEnableDelete
-    const Textinput = (
-        <View>
-            {/* textinput on/off */}
-            {writeDiary &&
-                <View>
-                    <StyledTextInput
-                        autoFocus={true}
-                        placeholder='Write a note'
-                        value={text}
-                        multiline={true}
-                        onChangeText={(text) => { setText(text); setReturnText(text) }}
-                    />
-                    <StyledHorizontallyAlignedItems
-                        style={{
-                            justifyContent: 'flex-end', alignItems: 'flex-end'
-                        }}
-                    >
-                        <TouchableOpacity onPress={() => { console.log("save the text"); setWriteDiary(false); setReturnText(text); }}
-                            style={{
-                                borderColor: '#8FDF70',
-                                borderWidth: 1,
-                                borderRadius: 50,
-                                backgroundColor: '#8FDF70',
-                                width: 50
-                            }}>
-                            {/* save a note */}
-                            <StyledButtonText>
-                                &#10003;
-                            </StyledButtonText>
-                        </TouchableOpacity>
-                        <TouchableOpacity onPress={() => { setWriteDiary(false); setText('') }}
-                            style={{
-                                borderColor: '#FF2511',
-                                borderWidth: 1,
-                                borderRadius: 50,
-                                backgroundColor: '#FF2511',
-                                width: 50
-                            }}>
-                            {/* delete a note     */}
-                            <StyledButtonText>
-                                &#65794;
-                            </StyledButtonText>
-                        </TouchableOpacity>
-                    </StyledHorizontallyAlignedItems>
-                </View>
-            }
-        </View >
-    )
-    return [(
+    const [showPlusButton, setShowPlusButton] = useState<boolean>(!writeDiaryEx && !enableDelete)
+    const updateShowPlus=(tf:boolean)=>{
+        setShowPlusButton(tf)
+    }
+    showPlusButtonEx=updateShowPlus
+    return (
         <View >
             {/* if user selects a note button, then stop displaying + button. */}
-            {enableDelete && 
-                <TouchableOpacity onPress={()=>{setEnableDelete(false)}}>
+            {enableDelete &&
+                <TouchableOpacity onPress={() => { setEnableDelete(false);updateShowPlus(true) }}>
                     <Text style={{
-                        color:'white'
+                        color: 'white'
                     }}>
                         Cancel
                     </Text>
                 </TouchableOpacity>
             }
-            {(!writeDiary && !enableDelete)  &&
+            {((!writeDiaryEx && !enableDelete)||showPlusButton) &&
                 <StyledCircleButton style={{
                     position: 'absolute',
                     // left:'50%',
@@ -393,60 +157,41 @@ function FunctionComponents(): [JSX.Element, JSX.Element, JSX.Element, string] {
                     {visible && //hidden buttons, visible when + button is clicked.
                         <View style={{
                             position: 'absolute',
-                            // left:'50%',
-                            width: '100%'
+                            width: '100%',
                         }}>
                             <Animated.View style={[{
                                 position: 'absolute',
                             }, animatedStyle1]}>
                                 <TouchableOpacity
-                                    onPress={openCamera}>
-                                    {/* open camera  */}
-                                    {/* &#x1F4F7; is the unicode for camera emoji */}
-                                    {/* <StyledText>
-                                        &#x1F4F7;
-                                    </StyledText> */}
+                                    onPress={() => { openCameraEx(); animationReset(); setVisible(false) }}>
                                     <Image source={icon_camera} key='camera' style={{ height: 50, width: 50, borderRadius: 50 }} />
-
                                 </TouchableOpacity>
                             </Animated.View>
                             <Animated.View style={[{
                                 position: 'absolute',
                             }, animatedStyle2]}>
-                                <TouchableOpacity onPress={openGallery}>
+                                <TouchableOpacity onPress={() => { openGalleryEx(); animationReset(); setVisible(false) }}>
                                     {/* open gallery */}
                                     <Image source={icon_gallery} key='gallery' style={{ height: 50, width: 50, borderRadius: 50 }} />
-
                                 </TouchableOpacity>
                             </Animated.View>
                             <Animated.View style={[{
                                 position: 'absolute',
                             }, animatedStyle3]}>
                                 <TouchableOpacity onPress={() => {
-                                    setWriteDiary(true); setVisible(false)
+                                    setWriteDiaryEx(true)
                                     animationReset()
-
+                                    setVisible(false)
                                 }}>
                                     <Image source={icon_note} key='note' style={{ height: 50, width: 50, borderRadius: 50 }} />
-
-                                </TouchableOpacity>
-                            </Animated.View>
-                            <Animated.View style={[{
-                                position: 'absolute',
-                            }, animatedStyle4]}>
-                                <TouchableOpacity onPress={removeCurrentDiary}>
-                                    {/* remove current diary */}
-                                    <Image source={icon_delete} key='delete' style={{ height: 50, width: 50, borderRadius: 50 }} />
-
                                 </TouchableOpacity>
                             </Animated.View>
                             <Animated.View style={[{
                                 position: 'absolute',
                             }, animatedStyle5]}>
-                                <TouchableOpacity onPress={()=>{setEnableDelete(true)}}>
+                                <TouchableOpacity onPress={() => { setEnableDelete(true); setVisible(false) }}>
                                     {/* enable edit mode */}
                                     <Image source={icon_edit} key='edit' style={{ height: 50, width: 50, borderRadius: 50 }} />
-
                                 </TouchableOpacity>
                             </Animated.View>
                         </View>
@@ -455,7 +200,6 @@ function FunctionComponents(): [JSX.Element, JSX.Element, JSX.Element, string] {
             }
 
         </View>
-    ), count != 0 ? <Photos content={photoContent as JSX.Element[]}></Photos> : <View ></View>,
-        Textinput, returnText]
+    )
 }
 export default FunctionComponents
